fix: HashVersioning by filtering fields
