--- packages/offix-server/README.md
@@ -1,5 +1,0 @@
-## Offix GraphQL Offline Server
-
-Documentation:
-
-https://offix.dev
