--- packages/offix-server/src/api/ObjectStateData.ts
@@ -1,9 +1,0 @@
-
-/**
- * ObjectStateData - defined type that abstract type of data
- * being handled by conflict resolution engine.
- * Data can take any form factor and contain different fields,
- * however it should always be used in unmodified state
- * (no new fields are added or removed)
- */
-export type ObjectStateData = any;
