--- packages/offix-server/src/states/HashObjectState.ts
@@ -1,50 +1,0 @@
-import { ConflictResolution } from "../api/ConflictResolution";
-import { ConflictResolutionStrategy } from "../api/ConflictResolutionStrategy";
-import { ObjectState } from "../api/ObjectState";
-import { ObjectStateData } from "../api/ObjectStateData";
-
-/**
- * Object state manager using a hashing method provided by user
- */
-export class HashObjectState implements ObjectState {
-  private hash: (object: any) => string;
-
-  constructor(hashImpl: (object: any) => string) {
-    this.hash = hashImpl;
-  }
-
-  public hasConflict(serverState: ObjectStateData, clientState: ObjectStateData) {
-    if (this.hash(serverState) !== this.hash(clientState)) {
-      return true;
-    }
-    return false;
-  }
-
-  public nextState(currentObjectState: ObjectStateData) {
-    // Hash can be calculated at any time and it is not added to object
-    return currentObjectState;
-  }
-
-  public resolveOnClient(serverState: ObjectStateData, clientState: ObjectStateData) {
-    return new ConflictResolution(false, serverState, clientState);
-  }
-
-  public reject(serverState: ObjectStateData, clientState: ObjectStateData) {
-    return new ConflictResolution(true, serverState, clientState);
-  }
-
-  public async resolveOnServer(strategy: ConflictResolutionStrategy,
-                               serverState: ObjectStateData,
-                               clientState: ObjectStateData) {
-    let resolvedState = strategy(serverState, clientState);
-
-    if (resolvedState instanceof Promise) {
-      resolvedState = await resolvedState;
-    }
-
-    resolvedState = this.nextState(resolvedState);
-
-    return new ConflictResolution(true, resolvedState, clientState);
-  }
-
-}
